Increment counter on config change
Increment a counter when a configuration variable is set or
deleted. When collected periodically (via collectd or such like)
and stored in graphite this data can be used to plot the times at
which configuration changes were made on a node. This should
make it easier for operators to relate configuration changes to
other metrics.

BugzID: 29704